Rename tree conversion folder
